# src/main/java/generated/com/panopticode/openapi/api/CookbookApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.18.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.panopticode.openapi.api;

import com.panopticode.openapi.model.ErrorModel;
import com.panopticode.openapi.model.Ingredient;
import com.panopticode.openapi.model.Recipe;
import com.panopticode.openapi.model.RecipeRef;
import com.panopticode.openapi.model.RequiredRecipeBody;
import java.util.UUID;
import com.panopticode.openapi.model.UpdateRecipeRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-28T03:09:42.850562636Z[Europe/London]", comments = "Generator version: 7.18.0")
@Validated
@Tag(name = "cookbook", description = "the cookbook API")
public interface CookbookApi {

    default CookbookApiDelegate getDelegate() {
        return new CookbookApiDelegate() {};
    }

    String PATH_ADD_INGREDIENT = "/recipes/{recipeId}/ingredients";
    /**
     * POST /recipes/{recipeId}/ingredients : Add new ingredient
     * Add a new ingredient to a recipe
     *
     * @param recipeId Recipe id (required)
     * @param ingredient Data object to save a new ingredient (required)
     * @return successful operation (status code 201)
     *         or Invalid input (status code 405)
     *         or Unexpected error (status code 200)
     */
    @Operation(
        operationId = "addIngredient",
        summary = "Add new ingredient",
        description = "Add a new ingredient to a recipe",
        tags = { "cookbook" },
        responses = {
            @ApiResponse(responseCode = "201", description = "successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Ingredient.class))
            }),
            @ApiResponse(responseCode = "405", description = "Invalid input", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "default", description = "Unexpected error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = CookbookApi.PATH_ADD_INGREDIENT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Ingredient> addIngredient(
        @NotNull @Parameter(name = "recipeId", description = "Recipe id", required = true, in = ParameterIn.PATH) @PathVariable("recipeId") UUID recipeId,
        @Parameter(name = "Ingredient", description = "Data object to save a new ingredient", required = true) @Valid @RequestBody Ingredient ingredient
    ) {
        return getDelegate().addIngredient(recipeId, ingredient);
    }


    String PATH_ADD_RECIPE = "/recipes";
    /**
     * POST /recipes : Add new recipe
     * Add a new recipe to the book
     *
     * @param requiredRecipeBody Data object to save a new recipe (required)
     * @return successful operation (status code 201)
     *         or Invalid input (status code 405)
     *         or Unexpected error (status code 200)
     */
    @Operation(
        operationId = "addRecipe",
        summary = "Add new recipe",
        description = "Add a new recipe to the book",
        tags = { "cookbook" },
        responses = {
            @ApiResponse(responseCode = "201", description = "successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Recipe.class))
            }),
            @ApiResponse(responseCode = "405", description = "Invalid input", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "default", description = "Unexpected error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = CookbookApi.PATH_ADD_RECIPE,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Recipe> addRecipe(
        @Parameter(name = "RequiredRecipeBody", description = "Data object to save a new recipe", required = true) @Valid @RequestBody RequiredRecipeBody requiredRecipeBody
    ) {
        return getDelegate().addRecipe(requiredRecipeBody);
    }


    String PATH_DELETE_ALL_INGREDIENTS = "/recipes/{recipeId}/ingredients";
    /**
     * DELETE /recipes/{recipeId}/ingredients : Deletes all ingredients
     * Deletes all ingredients off a recipe
     *
     * @param recipeId Recipe id (required)
     * @return Deletion completed successfully (status code 204)
     *         or Invalid recipe id (status code 404)
     *         or Unexpected error (status code 200)
     */
    @Operation(
        operationId = "deleteAllIngredients",
        summary = "Deletes all ingredients",
        description = "Deletes all ingredients off a recipe",
        tags = { "cookbook" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Deletion completed successfully"),
            @ApiResponse(responseCode = "404", description = "Invalid recipe id", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "default", description = "Unexpected error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = CookbookApi.PATH_DELETE_ALL_INGREDIENTS,
        produces = { "application/json" }
    )
    default ResponseEntity<Void> deleteAllIngredients(
        @NotNull @Parameter(name = "recipeId", description = "Recipe id", required = true, in = ParameterIn.PATH) @PathVariable("recipeId") UUID recipeId
    ) {
        return getDelegate().deleteAllIngredients(recipeId);
    }


    String PATH_DELETE_INGREDIENT = "/recipes/{recipeId}/ingredients/{ingredientId}";
    /**
     * DELETE /recipes/{recipeId}/ingredients/{ingredientId} : Deletes a ingredient
     * Deletes an existing ingredient off a recipe
     *
     * @param recipeId Recipe id (required)
     * @param ingredientId Ingredient id to delete (required)
     * @return Deletion completed successfully (status code 204)
     *         or Invalid recipe or ingredient id (status code 404)
     *         or Unexpected error (status code 200)
     */
    @Operation(
        operationId = "deleteIngredient",
        summary = "Deletes a ingredient",
        description = "Deletes an existing ingredient off a recipe",
        tags = { "cookbook" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Deletion completed successfully"),
            @ApiResponse(responseCode = "404", description = "Invalid recipe or ingredient id", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "default", description = "Unexpected error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = CookbookApi.PATH_DELETE_INGREDIENT,
        produces = { "application/json" }
    )
    default ResponseEntity<Void> deleteIngredient(
        @NotNull @Parameter(name = "recipeId", description = "Recipe id", required = true, in = ParameterIn.PATH) @PathVariable("recipeId") UUID recipeId,
        @NotNull @Parameter(name = "ingredientId", description = "Ingredient id to delete", required = true, in = ParameterIn.PATH) @PathVariable("ingredientId") UUID ingredientId
    ) {
        return getDelegate().deleteIngredient(recipeId, ingredientId);
    }


    String PATH_DELETE_RECIPE = "/recipes/{recipeId}";
    /**
     * DELETE /recipes/{recipeId} : Deletes a recipe
     * Deletes an existing recipe from the cookbook
     *
     * @param recipeId Recipe id to delete (required)
     * @return Deletion completed successfully (status code 204)
     *         or Invalid recipe id (status code 404)
     *         or Unexpected error (status code 200)
     */
    @Operation(
        operationId = "deleteRecipe",
        summary = "Deletes a recipe",
        description = "Deletes an existing recipe from the cookbook",
        tags = { "cookbook" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Deletion completed successfully"),
            @ApiResponse(responseCode = "404", description = "Invalid recipe id", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "default", description = "Unexpected error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = CookbookApi.PATH_DELETE_RECIPE,
        produces = { "application/json" }
    )
    default ResponseEntity<Void> deleteRecipe(
        @NotNull @Parameter(name = "recipeId", description = "Recipe id to delete", required = true, in = ParameterIn.PATH) @PathVariable("recipeId") UUID recipeId
    ) {
        return getDelegate().deleteRecipe(recipeId);
    }


    String PATH_GET_ALL_RECIPE_INGREDIENTS = "/recipes/{recipeId}/ingredients";
    /**
     * GET /recipes/{recipeId}/ingredients : Get all ingredients
     * Returns all ingredients related to a recipe
     *
     * @param recipeId Recipe id (required)
     * @return List of ingredients (status code 200)
     *         or Unexpected error (status code 200)
     */
    @Operation(
        operationId = "getAllRecipeIngredients",
        summary = "Get all ingredients",
        description = "Returns all ingredients related to a recipe",
        tags = { "cookbook" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of ingredients", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Ingredient.class)))
            }),
            @ApiResponse(responseCode = "default", description = "Unexpected error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = CookbookApi.PATH_GET_ALL_RECIPE_INGREDIENTS,
        produces = { "application/json" }
    )
    default ResponseEntity<List<Ingredient>> getAllRecipeIngredients(
        @NotNull @Parameter(name = "recipeId", description = "Recipe id", required = true, in = ParameterIn.PATH) @PathVariable("recipeId") UUID recipeId
    ) {
        return getDelegate().getAllRecipeIngredients(recipeId);
    }


    String PATH_GET_ALL_RECIPES = "/recipes";
    /**
     * GET /recipes : Get all recipes
     * Returns all recipes in the cookbook
     *
     * @return List of recipes (status code 200)
     *         or Unexpected error (status code 200)
     */
    @Operation(
        operationId = "getAllRecipes",
        summary = "Get all recipes",
        description = "Returns all recipes in the cookbook",
        tags = { "cookbook" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of recipes", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = RecipeRef.class)))
            }),
            @ApiResponse(responseCode = "default", description = "Unexpected error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = CookbookApi.PATH_GET_ALL_RECIPES,
        produces = { "application/json" }
    )
    default ResponseEntity<List<RecipeRef>> getAllRecipes(
        
    ) {
        return getDelegate().getAllRecipes();
    }


    String PATH_GET_RECIPE_BY_ID = "/recipes/{recipeId}";
    /**
     * GET /recipes/{recipeId} : Get a recipe
     * Returns a recipe by ID
     *
     * @param recipeId ID of the recipe to return (required)
     * @return Recipe (status code 200)
     *         or Invalid ID supplied (status code 400)
     *         or Recipe not found (status code 404)
     *         or Unexpected error (status code 200)
     */
    @Operation(
        operationId = "getRecipeById",
        summary = "Get a recipe",
        description = "Returns a recipe by ID",
        tags = { "cookbook" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Recipe", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Recipe.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid ID supplied", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "404", description = "Recipe not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "default", description = "Unexpected error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = CookbookApi.PATH_GET_RECIPE_BY_ID,
        produces = { "application/json" }
    )
    default ResponseEntity<Recipe> getRecipeById(
        @NotNull @Parameter(name = "recipeId", description = "ID of the recipe to return", required = true, in = ParameterIn.PATH) @PathVariable("recipeId") UUID recipeId
    ) {
        return getDelegate().getRecipeById(recipeId);
    }


    String PATH_UPDATE_INGREDIENT = "/recipes/{recipeId}/ingredients/{ingredientId}";
    /**
     * PUT /recipes/{recipeId}/ingredients/{ingredientId} : Update ingredient
     * Update an existing ingredient
     *
     * @param recipeId Recipe id (required)
     * @param ingredientId Ingredient id to update (required)
     * @return successful operation (status code 200)
     *         or Invalid ID supplied (status code 400)
     *         or Recipe or ingredient not found (status code 404)
     *         or Validation exception (status code 405)
     *         or Unexpected error (status code 200)
     */
    @Operation(
        operationId = "updateIngredient",
        summary = "Update ingredient",
        description = "Update an existing ingredient",
        tags = { "cookbook" },
        responses = {
            @ApiResponse(responseCode = "200", description = "successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Ingredient.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid ID supplied", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "404", description = "Recipe or ingredient not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "405", description = "Validation exception", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "default", description = "Unexpected error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = CookbookApi.PATH_UPDATE_INGREDIENT,
        produces = { "application/json" }
    )
    default ResponseEntity<Ingredient> updateIngredient(
        @NotNull @Parameter(name = "recipeId", description = "Recipe id", required = true, in = ParameterIn.PATH) @PathVariable("recipeId") UUID recipeId,
        @NotNull @Parameter(name = "ingredientId", description = "Ingredient id to update", required = true, in = ParameterIn.PATH) @PathVariable("ingredientId") UUID ingredientId
    ) {
        return getDelegate().updateIngredient(recipeId, ingredientId);
    }


    String PATH_UPDATE_RECIPE = "/recipes/{recipeId}";
    /**
     * PUT /recipes/{recipeId} : Update recipe
     * Update an existing recipe
     *
     * @param recipeId Recipe id to update (required)
     * @param updateRecipeRequest Data object to update an existing recipe (required)
     * @return successful operation (status code 200)
     *         or Invalid ID supplied (status code 400)
     *         or Recipe not found (status code 404)
     *         or Validation exception (status code 405)
     *         or Unexpected error (status code 200)
     */
    @Operation(
        operationId = "updateRecipe",
        summary = "Update recipe",
        description = "Update an existing recipe",
        tags = { "cookbook" },
        responses = {
            @ApiResponse(responseCode = "200", description = "successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RecipeRef.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid ID supplied", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "404", description = "Recipe not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "405", description = "Validation exception", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "default", description = "Unexpected error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = CookbookApi.PATH_UPDATE_RECIPE,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<RecipeRef> updateRecipe(
        @NotNull @Parameter(name = "recipeId", description = "Recipe id to update", required = true, in = ParameterIn.PATH) @PathVariable("recipeId") UUID recipeId,
        @Parameter(name = "UpdateRecipeRequest", description = "Data object to update an existing recipe", required = true) @Valid @RequestBody UpdateRecipeRequest updateRecipeRequest
    ) {
        return getDelegate().updateRecipe(recipeId, updateRecipeRequest);
    }

}